requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
rapidfuzz
imagehash
python-magic
opencv-python
pillow
paddleocr
paddlepaddle==3.0.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib